v1.1.0 - Added new feature to spesify quality and resolution
Added functions to be able to specify the quality and resolutions of downloaded media.

audioOrVideo function has an option to pick custom which forwards to customFormat. customFormat will allow the user to set the file extention (mp3, mp4, etc) and will then pass this onto customQuality. customQuality will then get the input for quality (best or worst), and if the user is downloading a video it will prompt for a resolution. This will then be passed back to the audioOrVideo function to get the download parameter.