Created a test for ffmpeg fallback
created a fallback if ffmpeg is not installed.

the program will loop and download both the audio and the video for the submitted url (fully separately).
If successfull it will then proced to warn the user and merge the two files using the  "moviepy" package.

moviepy isnt fast so I need to look into possible alternatives, however this does work