Created 2 potential fixes for ffmpeg
both methods download the audio and video separate, then merge them together using moviepy before iterating to the next video.

Method 1:  Uses the ydl option for playliststart and playlistend to iterates through each item in the playlist

Method 2: Scans the entered url and "scans" through each video collecting data such as the url. After this is complete the program will iterate through each of the collected urls and download.

Method 1 is prefered as its much faster when it comes to large playlists, as it doesnt need to scan through it twice (once to get data, twice to download)