updated packages
Updated to the newer versions of all packages apart from flet and moviepy.

flet was having an issue with the newest version, so it is one behind.
moviepy I think changed to using ffmpeg which is exactly the opposite of what I need. Therefore its stayed as the same version (for now possibly)